end2end-pipeline/services/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "FastAPI backend serving aggregated event analytics from ClickHouse."
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "clickhouse-connect==0.7.19",
  "pydantic==2.9.2",
  "structlog==24.4.0",
  "opentelemetry-api==1.27.0",
  "opentelemetry-sdk==1.27.0",
  "opentelemetry-exporter-otlp-proto-grpc==1.27.0",
  "opentelemetry-instrumentation-fastapi==0.48b0",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.3",
  "httpx==0.27.2",
  "ruff==0.6.9",
  "mypy==1.11.2",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]

[tool.mypy]
strict = true
python_version = "3.12"

[[tool.mypy.overrides]]
module = ["clickhouse_connect.*"]
ignore_missing_imports = true
implicit_reexport = true
